asl_cards/requirements.txt
# python == 3.6.1
pdfminer.six == 20170419
